wip: emulate better EXT_multisampled_render_to_texture
multisampled buffers are lost after resolve when using
EXT_multisampled_render_to_texture, but this wasn't the case with
our GL desktop emulation. We better simulate this, to match the
Android behavior and not rely on undefined behavior.

We also need to load the content of the resolved texture at
beginRenderPass time.